requirements.txt
langchain
langchain-community 
langchain-google-genai
google-generativeai
pypdf
pydantic
python-dotenv
